requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.27.0
psycopg2-binary==2.9.1
requests==2.28.1